refactor(4596): extract ClaudeE evidence stamps
Move the guarded durable stamp and handoff mutation into focused child modules so both parent files remain below the giant-file threshold without changing cleanup behavior.
